+++ src/app/components/example-collection.ts
@@ -48,6 +48,7 @@ import { TerraFormsExampleComponent } from './forms/terra-forms.example';
 import { RadioGroupComponentExample } from './forms/input/radio/example/radio-group.component.example';
 import { DataTableTemplateComponentExample } from './tables/data-table/example/data-table-template.component.example';
 import { FilterComponentExample } from './filter/example/filter.component.example';
+import { CheckboxGroupComponentExample } from './forms/checkbox-group/example/checkbox-group.component.example';
 
 export const examples:Array<Type<any>> = [
     TerraButtonComponentExample,
@@ -98,5 +99,6 @@ export const examples:Array<Type<any>> = [
     TerraFormsExampleComponent,
     RadioGroupComponentExample,
     DataTableTemplateComponentExample,
-    FilterComponentExample
+    FilterComponentExample,
+    CheckboxGroupComponentExample
 ];